requirements.txt
hydra-core>=1.3.2
omegaconf>=2.3.0
numpy>=1.24.0
plyfile>=1.0.0
scipy>=1.11.0
imageio[ffmpeg]>=2.31.0
tqdm>=4.66.0
taichi>=1.7.0
# Optional extras
opencv-python>=4.8.0
gsplat>=0.1.8
# torch>=2.2.0  # required for gsplat path; install matching your hardware build
